Added some unit tests
